Remove extra module within re_arrow_util
