+++ crates/store/re_chunk_store/src/writes.rs
@@ -384,7 +384,7 @@ impl ChunkStore {
                 });
             {
                 let is_semantically_empty =
-                    re_arrow_util::arrow_util::is_list_array_semantically_empty(list_array);
+                    re_arrow_util::is_list_array_semantically_empty(list_array);
 
                 column_metadata_state.is_semantically_empty &= is_semantically_empty;
             }
